refactor: decouple configuration loading and modularize worktree context
- Extract load_orchestra_config to vibe3.config.orchestra_settings and remove OrchestraConfig.from_settings.
- Move WorktreeContext to its own file to resolve environment module bloat/circularity.
- Consolidate webhook routing and signature verification logic into server/registry.py.
- Update all call sites and test suites to align with new modular structure.
