+++ src/vibe3/commands/status.py
@@ -6,6 +6,7 @@ from typing import Annotated, cast
 import typer
 
 from vibe3.commands.common import run_full_check_shortcut, trace_scope
+from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.models.flow import FlowStatusResponse
 from vibe3.models.orchestra_config import OrchestraConfig
 from vibe3.models.orchestration import IssueState
@@ -66,7 +67,7 @@ def status(
             run_full_check_shortcut()
 
         # 1. Orchestra State (Issues & Managers)
-        config = OrchestraConfig.from_settings()
+        config = load_orchestra_config()
         orch_snapshot = OrchestraStatusService.fetch_live_snapshot(config)
         snapshot_found = orch_snapshot is not None
