+++ src/vibe3/environment/__init__.py
@@ -5,7 +5,8 @@ from vibe3.environment.session import (
     SessionManager,
     TmuxSessionContext,
 )
-from vibe3.environment.worktree import WorktreeContext, WorktreeManager
+from vibe3.environment.worktree import WorktreeManager
+from vibe3.environment.worktree_context import WorktreeContext
 
 __all__ = [
     "WorktreeContext",
