+++ src/vibe3/execution/issue_role_sync_runner.py
@@ -8,6 +8,7 @@ from vibe3.agents.backends.codeagent import CodeagentBackend
 from vibe3.clients.git_client import GitClient
 from vibe3.clients.github_client import GitHubClient
 from vibe3.clients.sqlite_client import SQLiteClient
+from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.execution.actor_support import format_agent_actor
 from vibe3.execution.coordinator import ExecutionCoordinator
 from vibe3.execution.session_service import load_session_id
@@ -71,7 +72,7 @@ def run_issue_role_mode(
       `VIBE3_ASYNC_CHILD` only bypasses outer duplicate/capacity guards; the
       sync shell itself stays unchanged.
     """
-    config = OrchestraConfig.from_settings()
+    config = load_orchestra_config()
     issue = _load_issue_info(config, issue_number)
 
     store = SQLiteClient()
