+++ src/vibe3/server/app.py
@@ -1,28 +1,21 @@
 """vibe3 server - HTTP webhook receiver and CLI management."""
 
 import asyncio
-import hashlib
-import hmac
-import json
 import os
 import signal
 import sys
 from pathlib import Path
-from typing import Annotated, Any
+from typing import Annotated
 
 import typer
 import uvicorn
-from fastapi import APIRouter, Header, HTTPException, Request
-from fastapi.responses import JSONResponse
-from loguru import logger
 
 from vibe3.agents.backends.codeagent_config import find_missing_backend_commands
 from vibe3.clients.git_client import GitClient
+from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.models.orchestra_config import OrchestraConfig
 from vibe3.observability.logger import setup_logging
 from vibe3.orchestra.logging import orchestra_events_log_path, orchestra_log_dir
-from vibe3.runtime.heartbeat import HeartbeatServer
-from vibe3.runtime.service_protocol import GitHubEvent
 from vibe3.server.registry import (
     _build_server_with_launch_cwd,
     _kill_orchestra_tmux_session,
@@ -38,88 +31,6 @@ app = typer.Typer(
     help="Orchestra server: GitHub webhook receiver + heartbeat polling",
     no_args_is_help=True,
 )
-
-
-def _verify_signature(body: bytes, secret: str, header: str) -> bool:
-    """Return True if the HMAC-SHA256 signature matches."""
-    expected = "sha256=" + hmac.new(secret.encode(), body, hashlib.sha256).hexdigest()
-    return hmac.compare_digest(expected, header)
-
-
-def make_webhook_router(
-    heartbeat: HeartbeatServer,
-    webhook_secret: str | None,
-) -> APIRouter:
-    """Build a FastAPI router with GitHub webhook and health endpoints."""
-
-    router = APIRouter()
-
-    @router.post("/webhook/github")
-    async def receive_webhook(
-        request: Request,
-        x_github_event: str = Header(...),
-        x_hub_signature_256: str | None = Header(None),
-        x_github_delivery: str | None = Header(None),
-    ) -> JSONResponse:
-        body = await request.body()
-
-        if webhook_secret:
-            if not x_hub_signature_256:
-                raise HTTPException(status_code=401, detail="Missing webhook signature")
-            if not _verify_signature(body, webhook_secret, x_hub_signature_256):
-                raise HTTPException(status_code=403, detail="Invalid webhook signature")
-
-        try:
-            payload: dict[str, Any] = json.loads(body)
-        except json.JSONDecodeError as exc:
-            raise HTTPException(status_code=400, detail=f"Invalid JSON: {exc}") from exc
-
-        action = str(payload.get("action", ""))
-
-        logger.bind(
-            domain="orchestra",
-            action="webhook",
-            delivery=x_github_delivery,
-        ).info(
-            "Received: "
-            f"{x_github_event}/{action} "
-            f"(source=webhook, delivery={x_github_delivery or '-'})"
-        )
-
-        event = GitHubEvent(
-            event_type=x_github_event,
-            action=action,
-            payload=payload,
-            source="webhook",
-        )
-        await heartbeat.emit(event)
-
-        return JSONResponse({"status": "accepted", "event": x_github_event})
-
-    @router.get("/health")
-    async def health() -> JSONResponse:
-        return JSONResponse(
-            {
-                "status": "ok",
-                "services": heartbeat.service_names,
-                "queue_size": heartbeat.queue_size,
-            }
-        )
-
-    @router.get("/heartbeat")
-    async def heartbeat_status() -> JSONResponse:
-        """Legacy heartbeat status (use /status for full orchestra snapshot)."""
-        return JSONResponse(
-            {
-                "running": heartbeat.running,
-                "services": heartbeat.service_names,
-                "polling_interval": heartbeat.config.polling_interval,
-                "polling_enabled": heartbeat.config.polling.enabled,
-                "max_concurrent": heartbeat.config.max_concurrent_flows,
-            }
-        )
-
-    return router
 
 
 def _ensure_port_available(port: int) -> None:
@@ -278,7 +189,7 @@ def start(
         # Default and -v: show key events (tick completion, issue dispatch, etc.)
         _os.environ["VIBE3_ORCHESTRA_LOG_LEVEL"] = "INFO"
 
-    config = OrchestraConfig.from_settings()
+    config = load_orchestra_config()
     if not config.enabled:
         typer.echo("Orchestra is disabled in config (orchestra.enabled=false)")
         raise typer.Exit(1)
@@ -401,7 +312,7 @@ def start(
 @app.command()
 def status() -> None:
     """Show Orchestra server status."""
-    config = OrchestraConfig.from_settings()
+    config = load_orchestra_config()
     pid, is_valid = _validate_pid_file(config.pid_file)
 
     if pid is None:
@@ -430,7 +341,7 @@ def status() -> None:
 @app.command()
 def stop() -> None:
     """Stop Orchestra server via SIGTERM."""
-    config = OrchestraConfig.from_settings()
+    config = load_orchestra_config()
     pid_file = config.pid_file
     pid, is_valid = _validate_pid_file(pid_file)
