+++ tests/vibe3/models/test_orchestra_config.py
@@ -2,8 +2,8 @@
 
 from unittest.mock import patch
 
+from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.config.settings import VibeConfig
-from vibe3.models.orchestra_config import OrchestraConfig
 
 
 def test_from_settings_maps_supervisor_prompt_template() -> None:
@@ -11,6 +11,6 @@ def test_from_settings_maps_supervisor_prompt_template() -> None:
     settings.orchestra.supervisor_handoff.prompt_template = "orchestra.supervisor.apply"
 
     with patch("vibe3.config.settings.VibeConfig.get_defaults", return_value=settings):
-        config = OrchestraConfig.from_settings()
+        config = load_orchestra_config()
 
     assert config.supervisor_handoff.prompt_template == "orchestra.supervisor.apply"
